seccomp: add support for "clone3" syscall in default policy
If no seccomp policy is requested, then the built-in default policy in
containerd applies. This has no rule for "clone3" defined, nor any
default errno defined. So when runc receives the config it attempts to
determine a default errno, using logic defined in its commit:

  commit 7a8d7162f9d72f20d83eaa36aeb5426deecd58f2
  Author: Aleksa Sarai <cyphar@cyphar.com>
  Date:   Sun Jan 17 18:25:34 2021 +1100

    seccomp: prepend -ENOSYS stub to all filters

As explained in the above commit message, runc uses a heuristic to
decide which errno to return by default:

[quote]
  The solution applied here is to prepend a "stub" filter which returns
  -ENOSYS if the requested syscall has a larger syscall number than any
  syscall mentioned in the filter. The reason for this specific rule is
  that syscall numbers are (roughly) allocated sequentially and thus newer
  syscalls will (usually) have a larger syscall number -- thus causing our
  filters to produce -ENOSYS if the filter was written before the syscall
  existed.
[/quote]

Unfortunately clone3 appears to one of the edge cases that does not
result in use of ENOSYS, instead ending up with the historical EPERM
errno.

When glibc observes EPERM from clone3, it will not fallback to using
clone, instead treating it as a fatal error.

The clone3 syscall is much more complicated than clone, most notably its
flags are not exposed as a directly argument any more. Instead they are
hidden inside a struct. This means that seccomp filters are unable to
apply policy based on values seen in flags. Thus we can't directly
replicate the current "clone" filtering for "clone3". We can at least
ensure "clone3" returns ENOSYS errno, to trigger fallback to "clone"
at which point we can filter on flags.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>